Updated the neutron l3HA tool to use v3
The change adds Keystone v3 support to neutron l3 ha tool. This
change is needed to ensure that all l3-ha fail overs are
accomplished correctly.

The tool has been updated to skip / support routers that have
already bene setup using the built in neutron L3HA.

This is something that will need to be backported to Kilo

Change-Id: I94f21ff3fa511bc9da6b57cdf847719620818ece
Closes-Bug: #1499708
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

https://bugs.launchpad.net/openstack-ansible/+bug/1499708
https://review.openstack.org/#/c/229053/