Fix handling of cases where no L3 agents are alive
When no L3 agents were alive, the code would erroneously invoke
LOG.exception(), which could log an old exception which was harmless and
already caught and handled gracefully.  It would also still attempt to
perform migrations, which would lead to an IndexError when trying to
randomly pick which alive agent to migrate to:

    Traceback (most recent call last):
      File "/usr/bin/neutron-ha-tool", line 629, in <module>
        run(args)
      File "/usr/bin/neutron-ha-tool", line 122, in run
        l3_agent_migrate(qclient, args.noop, args.now)
      File "/usr/bin/neutron-ha-tool", line 314, in l3_agent_migrate
        target_id = random.choice(agent_alive_list)
      File "/usr/lib64/python2.7/random.py", line 275, in choice
        return seq[int(self.random() * len(seq))]  # raises IndexError if seq is empty
    IndexError: list index out of range

Instead we log an error and then give up on any migrations.