Fix handling of invalid agent to evacuate
When the agent to evacuate could not be found, the code would
erroneously invoke LOG.exception(), which could log an old exception
which was harmless and already caught and handled gracefully.  Worse, it
would also still attempt to retrieve the agent id, which would lead to:

    'NoneType' object has no attribute '__getitem__'

Instead we log an error and then give up on any migrations.