Make neutron-ha-tool recovery operations retry (bsc#965886)
If neutron-ha-tool's monitor action fails during a window in which the
database or message bus are down, or still starting up, its consequent
start action will fail.  There's no clean way to make Pacemaker do a
sensible number of retries at staggered intervals, so inside we embed
the retry logic inside the tool itself.  Ideally it would have gone in
the OCF RA, but it's uglier doing it in bash.

https://bugzilla.suse.com/show_bug.cgi?id=965886