neutron-ha-tool: filter agents based on agent_mode value (bsc#974495)
`target_agent_list` is filtering the list of agents excluding
the one that we want to migrate from.  We need to expand this
exclusion to the agents that contains a different configuration
agent_mode parameter.

For example, if we want to migrate from one `dvr_snat` agent to
another one, we need to exclude the `dvr`s one.

https://bugzilla.suse.com/show_bug.cgi?id=974495