neutron-ha-tool: make router migration synchronous by default
The new --wait-for-ports option (enabled by default) causes
neutron-ha-tool to wait for all not-distributed router-ports to be back
in the ACTIVE state on the target agent when migrating a router. This
can be used to avoid long work queues to pile up in the neutron agents
when migrating many routers in one go. And by that reduce the downtime
per migrated router. This behaviour can be turned off with by using
the --no-wait-for-ports option.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1012968