neutron-ha-tool: Add option to remove router namespace on remote nodes
The new --ssh-delete-namespace option will allow to have neutron-ha-tool
to delete the router's network namespace on the old node after it was
migrated to the new node. This is mainly indentend for evacuating
routers from nodes where the l3-agent is not running for some reason
(e.g. during the upgrade).