Refactor - create AgentPicker at setup time
We create the AgentPicker instances at setup time, only filling in the
agent list before the picking begins. This way users of the picker instances
are not responsible for creating instances, only setting agents and
then asking the picker to pick one. This way the internals and the
dependencies of the picker implementation is hidden from the users of
the picker instances.