Refactor - extract FakeNeutron, rename MockNeutronClient
MockNeutronClient was implementing both the neutron client's interface
and also providing methods for manipulating the underlying fake. To
communicate the fact that some method operates on the fake, they were
prefixed with tst_. This change extracts the neutron part from the
client, thus creating a very basic in-memory fake neutron. Also naming
these classes as fakes instead of mocks, as they are not asserting themselves.