[upgrade] Terminate processes before delete namespaces (bsc#1018582)
According to: https://lwn.net/Articles/580893/, when deleting a namespace
which still has processes running in it, the namespace will persist until all
processes in it exit. (Without it being visible to "ip netns list")

This commit reworks the code that deals with the namespace removal to first
try to kill all running processes (usually the neutron-metadata-proxy) in the
namespace before deleting the namespace itself.

To keep the code somewhat readable it introduces a separate class that handles
all the namespace removal related steps.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1018582