Add new option --exit-on-first-failure (bsc#1020914)
When the migration of a router fails, the current standard behaviour of
neutron-ha-tool is to continue migrating other routers and report the
number of failed migrations after finished migrating all routers of an
l3-agent. This commit adds the new argument "--exit-on-first-failure"
to allow to exit immediately after the first migration failure.

https://bugzilla.suse.com/show_bug.cgi?id=1020914