Add double router delete on router migration
On cloud6 environments with DVR enabled, the router is not properly
unbound from the L3 agent and requires a double delete to have it in a
clean state. This seems to be fixed from Mitaka version onward so this
workaround is needed for Liberty only. It should not break compatibility
with subsequent versions.

This is a workaround for bsc#1016943