Add infrastructure to handle TERM signal
neutron-ha-tool might be used as part of a service. It will be sent a
TERM signal upon service termination. This is an opportunity for
terminating ongoing operations in a safe manner, avoiding situations
when a router move (which is an add and a remove) would be interrupted
after a router being removed but not added back to the new agent.

This patch adds a signal handling mechanism that can protect some parts
of the code from being interrupted by the TERM signal.