Refactor - add MigrationResult
`migrate_router_safely` method returned a boolean that indicated the
success of the migration. Given that we want to put a check on agent
liveness to the method, it will be possible that the migration will be
skipped if the agent is alive, thus the return value needs to represent
that fact as well. By making the method return an object, multiple
return values can be added.